Extract common targets to mixin
Extract common targets to mixin
