+++ spec/models/manageiq/providers/azure/cloud_manager/targeted_refresher_spec.rb
@@ -45,14 +45,8 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh powered on VM" do
-        vm_resource_id = "#{@ems.subscription}\\#{@resource_group}\\microsoft.compute/virtualmachines\\#{@device_name}"
-
-        vm_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :vms,
-                                               :manager_ref => {:ems_ref => vm_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([vm_target], "_targeted/powered_on_vm_refresh")
+          refresh_with_cassette([vm_powered_on_target], "_targeted/powered_on_vm_refresh")
 
           assert_specific_az
           assert_specific_cloud_network
@@ -83,14 +77,8 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh powered off VM" do
-        vm_resource_id = "#{@ems.subscription}\\#{@resource_group}\\microsoft.compute/virtualmachines\\#{@vm_powered_off}"
-
-        vm_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :vms,
-                                               :manager_ref => {:ems_ref => vm_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([vm_target], "_targeted/powered_off_vm_refresh")
+          refresh_with_cassette([vm_powered_off_target], "_targeted/powered_off_vm_refresh")
 
           assert_specific_az
           assert_specific_flavor
@@ -118,14 +106,8 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh VM with managed disk" do
-        vm_resource_id = "#{@ems.subscription}\\#{@resource_group_managed_vm}\\microsoft.compute/virtualmachines\\#{@managed_vm}"
-
-        vm_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :vms,
-                                               :manager_ref => {:ems_ref => vm_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([vm_target], "_targeted/vm_with_managed_disk_refresh")
+          refresh_with_cassette([vm_with_managed_disk_target], "_targeted/vm_with_managed_disk_refresh")
 
           assert_specific_az
           assert_specific_flavor
@@ -154,79 +136,37 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh orchestration stack" do
-        stack_resource_id = "/subscriptions/#{@ems.subscription}/resourceGroups/miq-azure-test1/providers/Microsoft.Resources/deployments/spec-deployment-dont-delete"
-
-        stack_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                                  :association => :orchestration_stacks,
-                                                  :manager_ref => {:ems_ref => stack_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([stack_target], "_targeted/orchestration_stack_refresh")
+          refresh_with_cassette([parent_orchestration_stack_target], "_targeted/orchestration_stack_refresh")
 
           assert_stack_and_vm_targeted_refresh
         end
       end
 
       it "will refresh orchestration stack followed by Vm refresh" do
-        stack_resource_id = "/subscriptions/#{@ems.subscription}/resourceGroups/miq-azure-test1/providers/Microsoft.Resources/deployments/spec-deployment-dont-delete"
-
-        stack_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                                  :association => :orchestration_stacks,
-                                                  :manager_ref => {:ems_ref => stack_resource_id})
-
-        vm_resource_id = "#{@ems.subscription}\\miq-azure-test1\\microsoft.compute/virtualmachines\\spec0deply1vm0"
-        vm_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :vms,
-                                               :manager_ref => {:ems_ref => vm_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([stack_target], "_targeted/orchestration_stack_refresh")
+          refresh_with_cassette([parent_orchestration_stack_target], "_targeted/orchestration_stack_refresh")
 
           assert_stack_and_vm_targeted_refresh
 
-          refresh_with_cassette([vm_target], "_targeted/orchestration_stack_vm_refresh")
+          refresh_with_cassette([child_orchestration_stack_vm_target], "_targeted/orchestration_stack_vm_refresh")
           assert_stack_and_vm_targeted_refresh
         end
       end
 
       it "will refresh orchestration stack with vms" do
-        stack_resource_id = "/subscriptions/#{@ems.subscription}/resourceGroups/miq-azure-test1/providers/Microsoft.Resources/deployments/spec-deployment-dont-delete"
-
-        stack_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                                  :association => :orchestration_stacks,
-                                                  :manager_ref => {:ems_ref => stack_resource_id})
-
-        vm_resource_id1 = "#{@ems.subscription}\\miq-azure-test1\\microsoft.compute/virtualmachines\\spec0deply1vm0"
-        vm_target1      = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :vms,
-                                                     :manager_ref => {:ems_ref => vm_resource_id1})
-
-        vm_resource_id2 = "#{@ems.subscription}\\miq-azure-test1\\microsoft.compute/virtualmachines\\spec0deply1vm1"
-        vm_target2      = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :vms,
-                                                     :manager_ref => {:ems_ref => vm_resource_id2})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([stack_target, vm_target1, vm_target2], "_targeted/orchestration_stack_refresh")
+          refresh_with_cassette([parent_orchestration_stack_target,
+                                 child_orchestration_stack_vm_target,
+                                 child_orchestration_stack_vm_target2], "_targeted/orchestration_stack_refresh")
 
           assert_stack_and_vm_targeted_refresh
         end
       end
 
       it "will refresh orchestration stack followed by LoadBalancer refresh" do
-        stack_resource_id = "/subscriptions/#{@ems.subscription}/resourceGroups/miq-azure-test1/providers/Microsoft.Resources/deployments/spec-deployment-dont-delete"
-
-        stack_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                                  :association => :orchestration_stacks,
-                                                  :manager_ref => {:ems_ref => stack_resource_id})
-
-        lb_resource_id = "/subscriptions/#{@ems.subscription}/resourceGroups/miq-azure-test1/providers/Microsoft.Network/loadBalancers/spec0deply1lb"
-        lb_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :load_balancers,
-                                               :manager_ref => {:ems_ref => lb_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([stack_target], "_targeted/orchestration_stack_refresh")
+          refresh_with_cassette([parent_orchestration_stack_target], "_targeted/orchestration_stack_refresh")
 
           assert_stack_and_vm_targeted_refresh
 
@@ -236,11 +176,6 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh LoadBalancer created by stack" do
-        lb_resource_id = "/subscriptions/#{@ems.subscription}/resourceGroups/miq-azure-test1/providers/Microsoft.Network/loadBalancers/spec0deply1lb"
-        lb_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :load_balancers,
-                                               :manager_ref => {:ems_ref => lb_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
           refresh_with_cassette([lb_target], "_targeted/lb_created_by_stack_refresh")
 
@@ -255,14 +190,8 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh LoadBalancer" do
-        lb_resource_id = "/subscriptions/#{@ems.subscription}/"\
-                          "resourceGroups/miq-azure-test1/providers/Microsoft.Network/loadBalancers/rspec-lb1"
-        lb_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                               :association => :load_balancers,
-                                               :manager_ref => {:ems_ref => lb_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
-          refresh_with_cassette([lb_target], "_targeted/lb_refresh")
+          refresh_with_cassette([lb_non_stack_target], "_targeted/lb_refresh")
 
           assert_counts(
             :ext_management_system => 2,
@@ -317,47 +246,11 @@ describe ManageIQ::Providers::Azure::CloudManager::Refresher do
       end
 
       it "will refresh Template" do
-        template_resource_id = "https://miqazuretest14047.blob.core.windows.net/system/"\
-                               "Microsoft.Compute/Images/miq-test-container/"\
-                               "test-win2k12-img-osDisk.e17a95b0-f4fb-4196-93c5-0c8be7d5c536.vhd"
-
-        template_target = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :miq_templates,
-                                                     :manager_ref => {:ems_ref => template_resource_id})
-
         2.times do # Run twice to verify that a second run with existing data does not change anything
           refresh_with_cassette([template_target], "_targeted/template_refresh")
 
           assert_specific_template
         end
-      end
-
-      def lbs_targets
-        lb_resource_id1 = "/subscriptions/#{@ems.subscription}/"\
-                          "resourceGroups/miq-azure-test1/providers/Microsoft.Network/loadBalancers/rspec-lb1"
-        lb_target1      = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :load_balancers,
-                                                     :manager_ref => {:ems_ref => lb_resource_id1})
-
-        lb_resource_id2 = "/subscriptions/#{@ems.subscription}/"\
-                          "resourceGroups/miq-azure-test1/providers/Microsoft.Network/loadBalancers/rspec-lb2"
-        lb_target2      = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :load_balancers,
-                                                     :manager_ref => {:ems_ref => lb_resource_id2})
-        [lb_target1, lb_target2]
-      end
-
-      def lbs_vms_targets
-        vm_resource_id1 = "#{@ems.subscription}\\miq-azure-test1\\microsoft.compute/virtualmachines\\rspec-lb-a"
-        vm_target1      = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :vms,
-                                                     :manager_ref => {:ems_ref => vm_resource_id1})
-
-        vm_resource_id2 = "#{@ems.subscription}\\miq-azure-test1\\microsoft.compute/virtualmachines\\rspec-lb-b"
-        vm_target2      = ManagerRefresh::Target.new(:manager     => @ems,
-                                                     :association => :vms,
-                                                     :manager_ref => {:ems_ref => vm_resource_id2})
-        [vm_target1, vm_target2]
       end
 
       def assert_lbs_with_vms
